Add envId to the links and add middleware to deal with rewriting
