+++ pages/[envId]/[slug].tsx
@@ -2,14 +2,14 @@ import { GetStaticPaths, GetStaticProps } from "next";
 import { ParsedUrlQuery } from "querystring";
 import { FC } from "react";
 
-import { Content } from "../components/shared/Content";
-import { AppPage } from "../components/shared/ui/appPage";
-import { getDefaultMetadata, getItemByCodename, getSiteMenu } from "../lib/kontentClient";
-import { pageCodenames } from '../lib/routing';
-import { ValidCollectionCodename } from "../lib/types/perCollection";
-import { siteCodename } from "../lib/utils/env";
-import { createElementSmartLink, createFixedAddSmartLink } from "../lib/utils/smartLinkUtils";
-import { Block_Navigation, contentTypes, SEOMetadata, WSL_Page } from "../models";
+import { Content } from "../../components/shared/Content";
+import { AppPage } from "../../components/shared/ui/appPage";
+import { getDefaultMetadata, getItemByCodename, getSiteMenu } from "../../lib/kontentClient";
+import { pageCodenames } from '../../lib/routing';
+import { ValidCollectionCodename } from "../../lib/types/perCollection";
+import { siteCodename } from "../../lib/utils/env";
+import { createElementSmartLink, createFixedAddSmartLink } from "../../lib/utils/smartLinkUtils";
+import { Block_Navigation, contentTypes, SEOMetadata, WSL_Page } from "../../models";
 
 type Props = Readonly<{
   page: WSL_Page;
@@ -24,9 +24,7 @@ interface IParams extends ParsedUrlQuery {
 const pageCodenamesForGenericPrerender: ReadonlyArray<keyof typeof pageCodenames> = ['about-us'];
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  const paths = pageCodenamesForGenericPrerender.map(slug => (
-    { params: { slug } }
-  ))
+  const paths = Object.keys(pageCodenamesForGenericPrerender).map(slug => `/b0255462-358c-007b-0be0-43ee125ce1f0/${slug}`)
   return {
     paths,
     fallback: 'blocking',
