+++ pages/[envId]/articles/category/[category]/index.tsx
@@ -1,21 +1,21 @@
 import { ChevronDownIcon } from "@heroicons/react/24/solid";
-import { GetStaticProps } from "next";
+import { GetStaticPaths, GetStaticProps } from "next";
 import Link from "next/link";
 import { NextRouter, useRouter } from "next/router";
 import { FC, useState } from "react";
 
-import { ArticleItem } from "../../../../components/listingPage/ArticleItem";
-import { Content } from "../../../../components/shared/Content";
-import { useSiteCodename } from "../../../../components/shared/siteCodenameContext";
-import { AppPage } from "../../../../components/shared/ui/appPage";
-import { mainColorBgClass, mainColorBorderClass, mainColorHoverClass } from "../../../../lib/constants/colors";
-import { ArticlePageSize } from "../../../../lib/constants/paging";
-import { getArticlesCountByCategory, getArticlesForListing, getDefaultMetadata, getItemByCodename, getItemsTotalCount, getSiteMenu } from "../../../../lib/kontentClient";
-import { pageCodenames } from "../../../../lib/routing";
-import { ValidCollectionCodename } from "../../../../lib/types/perCollection";
-import { ArticleListingUrlQuery, ArticleTypeWithAll, categoryFilterSource, isArticleType } from "../../../../lib/utils/articlesListing";
-import { siteCodename } from "../../../../lib/utils/env";
-import { Article, Block_Navigation, SEOMetadata, taxonomies, WSL_Page } from "../../../../models";
+import { ArticleItem } from "../../../../../components/listingPage/ArticleItem";
+import { Content } from "../../../../../components/shared/Content";
+import { useSiteCodename } from "../../../../../components/shared/siteCodenameContext";
+import { AppPage } from "../../../../../components/shared/ui/appPage";
+import { mainColorBgClass, mainColorBorderClass, mainColorHoverClass } from "../../../../../lib/constants/colors";
+import { ArticlePageSize } from "../../../../../lib/constants/paging";
+import { getArticlesCountByCategory, getArticlesForListing, getDefaultMetadata, getItemByCodename, getItemsTotalCount, getSiteMenu } from "../../../../../lib/kontentClient";
+import { pageCodenames } from "../../../../../lib/routing";
+import { ValidCollectionCodename } from "../../../../../lib/types/perCollection";
+import { ArticleListingUrlQuery, ArticleTypeWithAll, categoryFilterSource, isArticleType } from "../../../../../lib/utils/articlesListing";
+import { siteCodename } from "../../../../../lib/utils/env";
+import { Article, Block_Navigation, SEOMetadata, taxonomies, WSL_Page } from "../../../../../models";
 
 
 type Props = Readonly<{
@@ -211,14 +211,14 @@ const ArticlesPage: FC<Props> = props => {
   )
 }
 
-export const getStaticPaths = async () => {
+export const getStaticPaths: GetStaticPaths = async () => {
 
   const getAllPagesForCategory = async (category: ArticleTypeWithAll) => {
     const totalCount = category === 'all' ? await getItemsTotalCount(false, 'article') : await getArticlesCountByCategory(false, category);
     const pagesNumber = Math.ceil((totalCount ?? 0) / ArticlePageSize);
     const pages = Array.from({ length: pagesNumber }).map((_, index) => index + 1);
     return pages.map(pageNumber => ({
-      params: { page: pageNumber.toString(), category },
+      params: { envId: '/b0255462-358c-007b-0be0-43ee125ce1f0/', page: pageNumber.toString(), category },
     }));
   };
